User logs into his account to reserve or pre-reserved a ticket.
It is an android application where user can create an account, log into
it and reserve a ticket to travel from an Egyptian city to another at a
specific time of his choice. User can return a pre-reserved ticket also
and get back his money corresponding to that ticket's price.